+++ Agenda de Contatos/src/dados/ContatoPessoal.java
@@ -1,6 +1,6 @@
-package entities;
+package dados;
 
-class ContatoPessoal extends Contato {
+public class ContatoPessoal extends Contato {
     private String endereco;
 
     public ContatoPessoal(String nome, String telefone, String email, String endereco) {
